+++ lnsubound2/lib/solve_mcf.h
@@ -2,4 +2,4 @@
 
 using namespace lemon;
 
-double solveMCF(SmartDigraph & g, SmartDigraph::ArcMap<int64_t> & cap, SmartDigraph::ArcMap<double> & cost, SmartDigraph::NodeMap<int64_t> & supplies, SmartDigraph::ArcMap<int64_t> & flow, int solverPar, SmartDigraph::NodeMap<double> & lnsPi);
+long double solveMCF(SmartDigraph & g, SmartDigraph::ArcMap<int64_t> & cap, SmartDigraph::ArcMap<long double> & cost, SmartDigraph::NodeMap<int64_t> & supplies, SmartDigraph::ArcMap<int64_t> & flow, int solverPar, SmartDigraph::NodeMap<long double> & lnsPi);
